+++ tests/unit/express/app.js
@@ -567,6 +567,17 @@ describe('App', function() {
             });
         });
 
+        describe('getHost', function() {
+            beforeEach(function() {
+                this.config = new Config({baseUrl: 'http://service.bistudio.com/root'});
+                this.app = this.appManager.buildApp(this.config, {name: 'getHost'});
+            });
+
+            it('should return base app location string (protocol + host)', function() {
+                this.app.getHost().should.be.equal('http://service.bistudio.com');
+            });
+        });
+
         describe('$buildExpressRouter', function() {
             it('should return new express Router object', function() {
                 var router = this.app.$buildExpressRouter();
@@ -660,6 +671,16 @@ describe('App', function() {
                     baseUrl: 'api.domain.com/root/path/',
                     routerUrl: '/user',
                     expectedBinding: '/root/path/user'
+                },
+                {
+                    baseUrl: 'http://api.domain.com/root/path/',
+                    routerUrl: '/user',
+                    expectedBinding: '/root/path/user'
+                },
+                {
+                    baseUrl: 'https://127.0.0.1:3000',
+                    routerUrl: '/user',
+                    expectedBinding: '/user'
                 },
                 {
                     baseUrl: '127.0.0.1:3000',