+++ tests/unit/mocks/config.js
@@ -5,10 +5,10 @@ module.exports = Config;
  * Config
  *
  */
-function Config() {
+function Config(data) {
     this.stores = {
         literal: {
-            store: {}
+            store: data || {}
         }
     };
 }
@@ -22,5 +22,6 @@ function Config() {
  * @return {undefined}
  */
 Config.prototype.get = function(key) {
+    return this.stores.literal.store[key];
 };